requirements.txt
amqp==5.1.1
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
backcall==0.2.0
billiard==4.1.0
celery==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
comm==0.1.3
cryptography==40.0.2
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.8.0
Django==3.2.15
django-allauth==0.54.0
django-debug-toolbar==4.1.0
django-extensions==3.2.3
django-redis==5.2.0
executing==1.2.0
flake8==6.0.0
idna==3.4
importlib-metadata==6.6.0
ipykernel==6.23.1
ipython==8.13.2
isort==5.12.0
jedi==0.18.2
# jupyter_client==8.2.0
# jupyter_core==5.3.0
kombu==5.3.0
matplotlib-inline==0.1.6
mccabe==0.7.0
nest-asyncio==1.5.6
netaddr==0.8.0
oauthlib==3.2.2
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2==2.9.6
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
Pygments==2.15.1
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2023.3
pywin32==306
pyzmq==25.1.0
redis==4.5.5
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
stripe==5.4.0
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.6.2
tzdata==2023.3
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
wrapt==1.15.0
# yookassa==2.3.6
zipp==3.15.0
